Switch UI system to Dear ImGui & related changes
This required switching from GLES2 to OpenGL 3+ with GL3W so let me know if it breaks anything.
This also required upgrading to a newer version of SDL_gpu.
  The setup script needs to be run again and it should install a newer version of SDL_gpu.
  Again, let me know if this doesn't work.
The "Debug Item" UI has been changed into a more generic "Cheats" UI, so more cheats can be added in the future.
The main menu now uses the temporary title logo (which will be changed once I come up with an actual name for the game).
The functionality for the chisel has been removed because it didn't work and I didn't really like how it used to work.
  A new way to precicely modify objects will probably be added at some point.